Changed AStar3D::Collection::find to use HashMap
Performance profiling with perf showed the aformentioned function as
the main bottleneck during AI turns. It used a linear search, which
turned out to be really expensive. The 'stl::multiset::find' function
couldn't be used because the search didn't compare the objects in
the set, but the object member 'node.h'.

This commit introduces a hash map which maps 'node.h' to its corresponding
'node'. Searches are now O(logN) instead of O(N).

Very unscientific testing (loading map 13 "blockade" and pressing "end
turn") gives a mere 24s for the AI turn instead of 36s before on a debug build
on a netbook, effectively cutting calculation time by 1/3.