Changed implementation of 'windbeeline' to constant time
The 'windbeeline' function is called a lot for every flying unit. Because
of it's O(N) complexity it consumed quite a bit of AI time, depending
on the number of airplanes. The new implementation is O(1) and barely
shows up on the performance profile.

It took great care that the new algorithm returns exactly the same as the
previous one for all values given.