Used 'memset' instead of a loop to clear temporary field markers
This patch touches a lot of code, but that's mostly replacing direct
access to a fields a.temp/a.temp2/tempw/temp3/temp4 with corresponding
getters and setters. The actual temporary markers are stored in arrays
on the gamemap, so that they use continuous memory and can be mass-
overwritten quickly with 'memset'.

I didn't really get why the union is there, but I replicated the
original behaviour.

On large maps, the speed-ups from this patch can be 10% or more for the
AI turn; it may help in responsivity on other places as well (didn't
measure it though).