Removed posDirs sparse array.
This one actually makes an AI a tad slower (not significantly though), but
removes some somewhat obscure code and uses quit a lot less memory by
saving the direction a path is traveled on on the node itself. This makes
for slightly worse memory access, but IMHO more understandable logic.