Removed the other sparse array (posHHops).
This commit removes the other sparse Array that saved height-changing
information and stores is on the node. Again, it doesn't make the code
faster, but IMHO clearer. There's still a bit of black magic going on
with this, and it would probably be better to store the information
in two vars, but for now I want to change as little as possible.