Introduced 'optimistic mode' for pathfinding.
While the standard A* algorithm is pretty good finding an optimal path
when the direct path is blocked, it has a bit of overhead for when there's
a direct path possible, like it is almost always the case for aircrafts.

In particular, in this case it is not neccessary to check the fields in
the vehicles back (we probably have checked them before anyways).

The direct path finder is very naive and never traces back a path when
it can't find a way forward. While that could easily possible with a
recursive finder, I doubt that it's worth the effort. We might find a
few more cases for optimistic mode with it, though.

In theory, if we didn't find a direct path in the first place, we could
re-check during regular pathfinding every now and then (like after height-
change) to see if there's a direct path between the current visited node
and the destination. In practice, I haven't found it worth the hassle.

It would be possible to be a bit more aggressive during optimistic mode
and only check the one or two directions towards the destination. I
haven't tried it yet, and I'm not sure it's a good idea with land
vehicles, but for sea and air units, where terrain is not relevant and
only hemming factors in into the movecost it should yield the optimal
path nonetheless.