Saving pointer to previous Node in Node.
That gets rid of the whole dir/HHops mess; for reconstructing the path
we simply trace the pointers back. For now I have eliminated the
'optimistic mode' because it really wasn't worth it, but I still plan to
implement a way for more aggressive pathfinding for aircrafts.

Then I implemented the VisitedContainer, which is a deque with a hmap
pointing to the Nodes therein. It's not a vector because a pointers to
a vector's elements are invalidated when the vector capacity grows -
that's not the case with deques, where pointers stay valid when
elements are pushed to it. That yields nearly no performance gains on
GCC 4.7, but that's an implementation detail - GCC 4.7 has a hash_map
implementation that has it's element as continuous in the memory as
possible, so iteration is quite fast, but that is not guaranteed. Boost
and 4.6 is slower, and so may be VS or other compilers.

Also, I changed the type of Path to a deque because it is constructed by
pushing things to the front, where a deque is slightly faster.