Changed the values of the hash_map of the VisitedContainer.
Previously, the hash_map stored pointers to the elements in the
VisitedContainer. However, there's no need to actually access a node
via hash map lookup. It's only ever used to 1) check if a field was
visited or 2) get the gval from that node. So we can just store the
gval in the hash map and be done with it. The 'VisitedContainer::find'
method was removed (as well as the 'fieldVisited' function as a wrapper
for it), and the methods 'contains' and 'gval' we introduced.